Corrigindo classes e realizando a persistencia no banco de dados

+++ SpringBoot/src/main/java/com/Cod3rs/SpringBoot/repositories/ProdutoRepository.java
@@ -4,6 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.Cod3rs.SpringBoot.models.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer>{
+public interface ProdutoRepository extends JpaRepository<Produto, Long>{
+
+	
 
 }
